Rename uflacsgenerator leftovers
